refactor: created models folder + models refactor
