+++ backend/app/models/DietaryFrequency.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+class DietaryFrequency(str, Enum):
+  daily = "diário"
+  couple_times_week = "algumas vezes na semana"
+  rarely = "raramente"
+  never = "nunca"
