+++ backend/app/models/FormData.py
@@ -1,31 +1,19 @@
-# app/models.py
 from enum import Enum
 from pydantic import BaseModel
-
-class SexOptions(str, Enum):
-  heterosexual_male = "Homem hetero"
-  heterosexual_female = "Mulher hetero"
-  trans_male = "Homem trans"
-  trans_female = "Mulher trans"
-  other = "Outro"
-
-class DietaryFrequency(str, Enum):
-  daily = "diário"
-  couple_times_week = "algumas vezes na semana"
-  rarely = "raramente"
-  never = "nunca"
+from GenderOptions import GenderOptions
+from DietaryFrequency import DietaryFrequency
 
 class FormData(BaseModel):
-  # Section 1: General User Data
+  #Section 1: General User Data
   age: int
-  sex: SexOptions
+  sex: GenderOptions
   weight_kg: float
   height_cm: int
   state: str = ""
   city: str = ""
   family_medical_history: str = ""
   
-  # Section 2: Dietary Habits
+  #Section 2: Dietary Habits
   fruits: DietaryFrequency
   vegetables: DietaryFrequency
   grains: DietaryFrequency
@@ -35,13 +23,13 @@ class FormData(BaseModel):
   ultra_processed_foods: DietaryFrequency
   sweets: DietaryFrequency
   water_intake_liters: float
-  special_dietary_practices: str | None = None
+  special_dietary_practices: str = ""
   breakfast_time: str = ""
   lunch_time: str = ""
   dinner_time: str = ""
   snack_time: str = ""
   
-  # Section 3: Symptoms
+  #Section 3: Symptoms
   fatigue: bool
   hair_loss: bool
   dry_skin: bool
@@ -53,24 +41,24 @@ class FormData(BaseModel):
   difficulty_concentrating: bool
   other_symptoms: str = ""
   
-  # Section 4: Medication
+  #Section 4: Medication
   regular_medication_use: bool
   medications_list: str = ""
   taking_supplements: bool
   supplements_list: str = ""
   frequency_dosage: str = ""
   
-  # Section 5: Lifestyle
+  #Section 5: Lifestyle
   physical_activity_frequency: str
   sleep_hours_per_night: int
   perceived_stress_level: str
   
-  # Section 6: Preferences
+  #Section 6: Preferences
   favorite_foods: str = ""
   avoided_foods: str
   dietary_restrictions: str
   
-  # Section 7-9: Goals/Consent/Notes
+  #Section 7-9: Goals/Consent/Notes
   nutritional_goal: str = ""
   consent_given: bool
   additional_notes: str = ""
